Add proteins gff as pipeline input

+++ modules/local/plot_contig_map/main.nf
@@ -8,27 +8,30 @@ process PLOT_CONTIG_MAP {
       tuple val(meta), val(set_name), path(tab)
     
     output:
-      tuple val(meta), val(set_name), path("${set_name}_mapping_results"), path("${set_name}_prot_ann_table_filtered.tsv")
+      tuple val(meta), val(set_name), path("${set_name}_mapping_results"), path("${set_name}_prot_ann_table_filtered.tsv"), optional: true
     
     script:
     """
-  	# get only contig IDs that have at least one annotation hit
-    cat ${tab} | sed 's/|/VIRIFY/g' > virify.tmp 
-	  export IDS=\$(awk 'BEGIN{FS="\\t"};{if(\$6!="No hit"){print \$1}}' virify.tmp | sort | uniq | grep -v Contig)
-	  head -1 ${tab} > ${set_name}_prot_ann_table_filtered.tsv
-	  for ID in \$IDS; do
-		  awk -v id="\$ID" '{if(id==\$1){print \$0}}' virify.tmp >> ${set_name}_prot_ann_table_filtered.tsv
-	  done
-    sed -i 's/VIRIFY/|/g' ${set_name}_prot_ann_table_filtered.tsv
-    mkdir -p ${set_name}_mapping_results
-    cp ${set_name}_prot_ann_table_filtered.tsv ${set_name}_mapping_results/
-    make_viral_contig_map.R -o ${set_name}_mapping_results -t ${set_name}_prot_ann_table_filtered.tsv
-    
-    # add pdfs into one folder and compress it
-    mkdir -p ${set_name}_mapping_results/plot_pdfs
-    mv ${set_name}_mapping_results/*.pdf ${set_name}_mapping_results/plot_pdfs/
-    tar czf ${set_name}_mapping_results/plot_pdfs.tar.gz ${set_name}_mapping_results/plot_pdfs
-    rm -rf ${set_name}_mapping_results/plot_pdfs
+    # if input has more than 1 line (which is header)
+    if (( \$(wc -l < ${tab}) > 1 )); then
+        # get only contig IDs that have at least one annotation hit
+        cat ${tab} | sed 's/|/VIRIFY/g' > virify.tmp 
+            export IDS=\$(awk 'BEGIN{FS="\\t"};{if(\$6!="No hit"){print \$1}}' virify.tmp | sort | uniq | grep -v Contig)
+            head -1 ${tab} > ${set_name}_prot_ann_table_filtered.tsv
+            for ID in \$IDS; do
+                awk -v id="\$ID" '{if(id==\$1){print \$0}}' virify.tmp >> ${set_name}_prot_ann_table_filtered.tsv
+            done
+        sed -i 's/VIRIFY/|/g' ${set_name}_prot_ann_table_filtered.tsv
+        mkdir -p ${set_name}_mapping_results
+        cp ${set_name}_prot_ann_table_filtered.tsv ${set_name}_mapping_results/
+        make_viral_contig_map.R -o ${set_name}_mapping_results -t ${set_name}_prot_ann_table_filtered.tsv
+        
+        # add pdfs into one folder and compress it
+        mkdir -p ${set_name}_mapping_results/plot_pdfs
+        mv ${set_name}_mapping_results/*.pdf ${set_name}_mapping_results/plot_pdfs/
+        tar czf ${set_name}_mapping_results/plot_pdfs.tar.gz ${set_name}_mapping_results/plot_pdfs
+        rm -rf ${set_name}_mapping_results/plot_pdfs
+    fi
     """
 }
